refactor: move analytics as a service
