+++ src/controllers/wordless.js
@@ -1,4 +1,4 @@
-import GA from "../modules/analytics.js";
+import GA from "../services/analytics.js";
 import Game from "../modules/game.js";
 import UI from "../modules/ui.js";
 import Input, { INPUT_CHANGE_EVENT, INPUT_SUBMIT_EVENT } from "./input.js";
@@ -139,12 +139,18 @@ const stopLoading = () => {
  * Registers the event handlers for user input and initializes the game.
  */
 const start = async () => {
+  // load UI components
+  UI.init();
+
+  // load Input controller
   Input.init();
 
+  // handle input change event
   document.addEventListener(INPUT_CHANGE_EVENT, (event) => {
     handleOnInputChange(event.detail);
   });
 
+  // handle input submit event
   document.addEventListener(INPUT_SUBMIT_EVENT, (event) => {
     handleOnInputSubmit(event.detail);
   });
